+++ client/src/Pages/Modals/Template.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "../../CSS/TemplateModal.css";
+import "../../css/TemplateModal.css";
 
 export default function Template(props) {
   if (!props.status) {
@@ -20,19 +20,40 @@ export default function Template(props) {
         <div className="row">
           <img
             id="modal-add-button"
-            src="image/template/sample.png"
+            src="images/template/sample.png"
+            alt="sampleImage"
             onClick={
               // TODO : 누르면 템플릿 적용되도록 구현
               () => props.onClose()
             }
           />
-          <img id="modal-add-button" src="image/template/sample.png" />
-          <img id="modal-add-button" src="image/template/sample.png" />
+          <img
+            id="modal-add-button"
+            src="images/template/sample.png"
+            alt="sampleImage"
+          />
+          <img
+            id="modal-add-button"
+            src="images/template/sample.png"
+            alt="sampleImage"
+          />
         </div>
         <div className="row">
-          <img id="modal-add-button" src="image/template/sample.png" />
-          <img id="modal-add-button" src="image/template/sample.png" />
-          <img id="modal-add-button" src="image/template/sample.png" />
+          <img
+            id="modal-add-button"
+            src="images/template/sample.png"
+            alt="sampleImage"
+          />
+          <img
+            id="modal-add-button"
+            src="images/template/sample.png"
+            alt="sampleImage"
+          />
+          <img
+            id="modal-add-button"
+            src="images/template/sample.png"
+            alt="sampleImage"
+          />
         </div>
       </div>
     </div>
